refactor(blackjack): Add the test for players
